requirements.txt
pygame >= 1.9.6
ruamel.yaml >= 0.16